fix: update chart options to chart container - reduces code redundancy
